Renamed tabstickers component, changed sticker sizing to em from, split stickers into separate components.
