+++ assets/js/apps/search/components/TabStickers/DatasetTabSticker.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import TabSticker from './TabSticker';
+
+
+export const DatasetTabSticker = () => {
+    return (<TabSticker initials="DS"></TabSticker>);
+};
+
+export default DatasetTabSticker;
