+++ assets/js/apps/search/components/TabStickers/ExperimentTabSticker.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import TabSticker from './TabSticker';
+export const ExperimentTabSticker = () => {
+    return (<TabSticker initials="E"></TabSticker>);
+};
+
+
+export default ExperimentTabSticker;
