+++ assets/js/apps/search/components/TabStickers/ProjectTabSticker.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import TabSticker from './TabSticker';
+export const ProjectTabSticker = () => {
+    return (<TabSticker initials="P"></TabSticker>);
+};
+
+
+export default ProjectTabSticker;
