+++ assets/js/apps/search/components/TabStickers/TabSticker.jsx
@@ -37,3 +37,5 @@ export default function TabSticker(props) {
         </div>
     )
 }
+
+
